move combine_keys to utils
